feat: implemented logic to extract and print last row and last column from 2D matrix

+++ Introduction to Programming/Final Exam/Find_the_Missing_Number.c
@@ -7,20 +7,20 @@ int main()
 
     for (int i = 0; i < T; i++)
     {
-        int arr[4];
+        long long arr[4];
 
         for (int j = 0; j < 4; j++)
         {
-            scanf("%d", &arr[j]);
+            scanf("%lld", &arr[j]);
         }
 
-        int M = arr[0];
-        int A = arr[1];
-        int B = arr[2];
-        int C = arr[3];
+        long long M = arr[0];
+        long long A = arr[1];
+        long long B = arr[2];
+        long long C = arr[3];
 
         if (M % (A * B * C) == 0)
-            printf("%d", M / (A * B * C));
+            printf("%lld", M / (A * B * C));
         else printf("%d", -1);
 
         printf("\n");
